refactor: move size_limit fit logic into _CacheDatasource.
