+++ pytoda/datasets/_csv_lazy_dataset.py
@@ -1,6 +1,4 @@
 """Implementation of _CsvLazyDataset."""
-import os
-import warnings
 from collections import OrderedDict
 
 import numpy as np
@@ -9,15 +7,6 @@ from ._cache_datasource import _CacheDatasource
 from .base_dataset import KeyDataset
 from ._csv_statistics import _CsvStatistics
 from ..types import FeatureList, Hashable
-
-
-def sizeof_fmt(num, suffix='B'):
-    """Source: https://stackoverflow.com/a/1094933"""
-    for unit in ['', 'Ki', 'Mi', 'Gi', 'Ti', 'Pi', 'Ei', 'Zi']:
-        if abs(num) < 1024.0:
-            return "%3.1f%s%s" % (num, unit, suffix)
-        num /= 1024.0
-    return "%.1f%s%s" % (num, 'Yi', suffix)
 
 
 class _CsvLazyDataset(KeyDataset, _CacheDatasource, _CsvStatistics):
@@ -48,17 +37,7 @@ class _CsvLazyDataset(KeyDataset, _CacheDatasource, _CsvStatistics):
         """
         self.chunk_size = chunk_size
 
-        size_limit = 1073741824  # default limit of 1GiB from diskcash
-        file_size = os.path.getsize(filepath)
-        if file_size > size_limit:
-            size_limit = file_size
-            message = (
-                f'Temporary directory for caching can be up to {size_limit} '
-                f'bytes ({sizeof_fmt(size_limit)}) large to fit data.'
-            )
-            # ResourceWarning is usually filtered by default
-            warnings.warn(message, ResourceWarning)
-        _CacheDatasource.__init__(self)
+        _CacheDatasource.__init__(self, fit_size_limit_filepath=filepath)
         _CsvStatistics.__init__(
             self, filepath, feature_list=feature_list, **kwargs
         )  # calls setup_datasource
